+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - 所有権インターフェースの汎用性拡張
+- Chevre転送取引へのignorePaymentCard指定を削除
 
 ### Deprecated